Llama/requirements.txt
streamlit>=1.31.0
sentence-transformers>=3.0.1
pinecone-client>=5.0.1
pandas>=2.2.0
openai>=1.12.0
transformers>=4.0.0
torch>=1.9.0
huggingface_hub>=0.0.12